requirements_prod.txt
python-dotenv==1.0.1
SQLAlchemy==2.0.27
pymysql==1.1.0
mysql-connector-python==8.3.0
requests~=2.32.3
boto3~=1.34.110